+++ CHANGELOG.md
@@ -7,7 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
-- Prop `changeOnHover` to `product-summary-image`.
+- Prop `hoverImageLabel` to `product-summary-image`.
+- Created function to normalize product-summary and use it uniformly in other components.
 
 ## [2.31.0] - 2019-08-01
 ### Added